Atualizando códigos

+++ TUR/ErusBot/Motors.h
@@ -2,6 +2,6 @@
 #define MOTORS_H
 
 void setupMotors(void);
-void setMotor(int motor, uint8_t velocity, uint8_t direction);
+void setMotor(int motor, int velocity, uint8_t direction);
 
 #endif
